Move HandleTableWalker into LegacyRuntime
